backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.1.0
psycopg2-binary
python-dotenv
httpx
pydantic
sqlparse
pandas
typing-extensions
google-generativeai==0.3.2"aiohttp==3.9.3" 
aiohttp==3.9.3 